Specifications
This is an experiment of Automatic Compliance check.
The large language models applied: GPT-3 fine-tuned models and GPT-3.5 models.
Because extra expending would be generated, the OpenAi key of fine-tuned models cannot be provided. 
However, the original dataset, datasets for fine-tuning, validation, and unifying tests are all provided to help researchers replicate the experiment.